Remove .eslintrc.cjs and replace the react icon with a fill color
Changes were introduced to improve code readability and app appearance. The .eslintrc.cjs was removed as it was no longer needed due to our migration to a config-file-free approach with Vite. Also improved the appearance of the react icon in the NavBar by updating the fill color.

More files were renamed according to the newly adopted naming convention.
The application now also supports SVGs thanks to the added vite-plugin-svgr in the package.json.
Firebase was added to the project to enable future authentication functionality, hence, a Login component was created, though it currently doesn't include any business logic.