Remove AuthNavbar and refine UI components
Deleted the AuthNavbar component as it is no longer being used. Updated the ProductCardList style to improve layout alignment. Added autoAnimate to the innerBody of DefaultNavbar to enhance user interaction and visual appeal. Adjusted the ProductCard's width and padding for better aesthetics. Lastly, extended the width of the DefaultNavbar and added an app-container for better site-wide consistency and layout. This commits aims to improve overall user experience and interface consistency.