Update product card list layout and fetch method
Removed the delay in fetching the products for productCardList component. The fetch is now immediate as the delay was not needed. Adjusted the style of product card list to provide better spacing and alignment. Changes improve the visual part of the product, and improve user experience by not waiting for unnecessary delay. Products are now arranged with 'space-around' to give them equal space around them. Overflow removed from card styles for better aspect ratio consistency.