Add Firebase authentication and user creation
In this commit, we've implemented Google Sign-In using Firebase Authentication. We've added a method to create a Firestore document for each signed-in user if it doesn't exist already. Additionally, we refactored some routing links and updated our package dependencies.

Why:
As part of delivering a seamless login experience to our users, we introduced Google Sign-In functionality. Storing user details on Firestore enhances our capacity to personalize their experience based on their profile data. Further, the clean-up on routing links provides manageable and human-readable paths that adhere to best coding practices.