"Add Carousel to Product Card List and enhance styling for responsiveness
Updated the Product Card List component to use a Carousel layout for better user experience. This layout is more mobile-friendly and easier to interact with on smaller screens. To do this, the 'primereact' library was added to the project, the number of items displayed was updated, and responsiveness settings were configured.

Additional enhancements were made to the overall application styling, including changes in product card sizes depending on screen resolution, transition effects, and updates in main app renders.

New dependencies were also added to package.json for the PrimeReact library, bootstrap, primeicons, and react-bootstrap. These will be installed on the next npm install command. Some existing packages are updated as well. Notably, the product card now has a DeferredContent component to defer loading of images until it's required."